histopath-mets-cnn-vit-mil/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "histopath-mets-cnn-vit-mil"
version = "0.1.0"
description = "Histopathology Metastasis Detection using CNN, ViT, and MIL"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Histopath Team", email = "team@histopath.ai"}
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pytorch-lightning>=2.1.0",
    "timm>=0.9.7",
    "hydra-core>=1.3.2",
    "mlflow>=2.7.1",
    "numpy>=1.24.3",
    "pandas>=2.0.3",
    "scikit-learn>=1.3.0",
    "opencv-python>=4.8.1",
    "Pillow>=10.0.0",
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "openslide-python>=1.3.0",
    "albumentations>=1.3.1",
    "shap>=0.42.1",
    "captum>=0.6.0",
    "pytorch-grad-cam>=1.4.8",
    "streamlit>=1.25.0",
    "fastapi>=0.103.1",
    "uvicorn[standard]>=0.23.2",
    "pydantic>=2.3.0",
    "tqdm>=4.65.0",
    "rich>=13.5.2",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
histopath-train = "histopath.train.train:main"
histopath-eval = "histopath.eval.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["histopath"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true